etl/requirements.txt
backoff==2.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
elastic-transport==8.13.1
elasticsearch==8.14.0
idna==3.7
psycopg2-binary==2.9.9
pydentic==0.0.1.dev3
python-dotenv==1.0.1
python-stdnum==1.20
requests==2.32.3
urllib3==2.2.2
